refactor: share World's list copy between entities and lore
The two copy() builders repeated the lowercase rule and five of seven fields; worldListCopy takes the kind's translated search scope and empty subtext.
